fixed merging conflict, updated Bookmark.jsx, removed BookmarkHeader.jsx and Sidebar.jsx. Integrated sidebar in book
